fix proxy methods tests
