+++ test/util/test.test.js
@@ -1,0 +1,1 @@
+// todo add tests for all util functions
